Added test to test_City_Grid.cpp. Tests getHouseAdjacent() with houses on the edges of the grid.

+++ src/Simulator_Basic_A.cpp
@@ -130,7 +130,7 @@ void Simulator_Basic_A::moveResident (Resident* res, int numOfTries)
     sort(housesInRange.begin(), housesInRange.end(), compare<const House*>);
 
     RandomIntegers randomIntegers{};
-    int quantity = min(housesInRange.size(), numOfTries);
+    int quantity = min((int)housesInRange.size(), numOfTries);
     vector<int> randIndices = randomIntegers.getRandomIntegers(housesInRange.size(), quantity);
     for (size_t ii=0; ii<randIndices.size(); ++ii)
     {
